wayland: move the output capture source code into it's own file
