+++ src/server/frontend_wayland/ext_image_capture_v1.h
@@ -32,14 +32,11 @@
 namespace mir
 {
 class Executor;
-namespace compositor { class ScreenShooterFactory; }
 namespace input { class CursorObserverMultiplexer; }
 namespace renderer::software { class RWMappable; }
 namespace time { class Clock; }
 namespace frontend
 {
-class SurfaceStack;
-
 class ExtImageCopyCaptureSessionV1;
 
 class ExtImageCopyBackend
@@ -121,13 +118,6 @@ private:
 
     std::shared_ptr<ExtImageCopyBackend> backend;
 };
-
-auto create_ext_output_image_capture_source_manager_v1(
-    wl_display* display,
-    std::shared_ptr<Executor> const& wayland_executor,
-    std::shared_ptr<compositor::ScreenShooterFactory> const& screen_shooter_factory,
-    std::shared_ptr<SurfaceStack> const& surface_stack)
-    -> std::shared_ptr<wayland::OutputImageCaptureSourceManagerV1::Global>;
 
 auto create_ext_image_copy_capture_manager_v1(
     wl_display* display,
